Moving Dwarf dump to extras so LAVA can use it
